revert changes in `@uppy/progress-bar`
