--- packages/@uppy/progress-bar/.npmignore
@@ -1,2 +1,0 @@
-
-tsconfig.*
