--- packages/@uppy/progress-bar/src/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProgressBar.tsx'
