--- packages/@uppy/progress-bar/tsconfig.build.json
@@ -1,35 +1,0 @@
-{
-  "extends": "../../../tsconfig.shared",
-  "compilerOptions": {
-    "noImplicitAny": false,
-    "outDir": "./lib",
-    "paths": {
-      "@uppy/utils/lib/*": [
-        "../utils/src/*"
-      ],
-      "@uppy/core": [
-        "../core/src/index.js"
-      ],
-      "@uppy/core/lib/*": [
-        "../core/src/*"
-      ]
-    },
-    "resolveJsonModule": false,
-    "rootDir": "./src",
-    "skipLibCheck": true
-  },
-  "include": [
-    "./src/**/*.*"
-  ],
-  "exclude": [
-    "./src/**/*.test.ts"
-  ],
-  "references": [
-    {
-      "path": "../utils/tsconfig.build.json"
-    },
-    {
-      "path": "../core/tsconfig.build.json"
-    }
-  ]
-}
