--- packages/@uppy/progress-bar/tsconfig.json
@@ -1,30 +1,0 @@
-{
-  "extends": "../../../tsconfig.shared",
-  "compilerOptions": {
-    "emitDeclarationOnly": false,
-    "noEmit": true,
-    "paths": {
-      "@uppy/utils/lib/*": [
-        "../utils/src/*"
-      ],
-      "@uppy/core": [
-        "../core/src/index.js"
-      ],
-      "@uppy/core/lib/*": [
-        "../core/src/*"
-      ]
-    }
-  },
-  "include": [
-    "./package.json",
-    "./src/**/*.*"
-  ],
-  "references": [
-    {
-      "path": "../utils/tsconfig.build.json"
-    },
-    {
-      "path": "../core/tsconfig.build.json"
-    }
-  ]
-}
